requirements.txt
discord.py @ git+https://github.com/robin-weaver/discord.py-for-selfbot
websockets
snscrape